+++ tt_metal/fabric/ccl/ccl_common.cpp
@@ -6,6 +6,7 @@
 
 #include <cstdint>
 
+#include <tt-metalium/experimental/fabric/fabric.hpp>
 #include <tt_metal/fabric/erisc_datamover_builder.hpp>
 #include <tt-metalium/hal.hpp>
 #include <tt-metalium/host_api.hpp>
